refactor: Se actualizaron las migraciones y algunos modelos para que este mas estandarizado y sea mantenible

+++ database/migrations/2024_01_15_091523_create_plan_pago_respaldo_table.php
@@ -14,32 +14,41 @@ return new class extends Migration
         Schema::create('plan_pago_respaldo', function (Blueprint $table) {
             $table->id();
 
-            // Datos del plan_pago original
+            // --- Datos copiados del plan_pago original ---
             $table->date('fecha_inicio')->nullable();
             $table->date('fecha_ultima_amortizacion')->nullable();
             $table->date('fecha_fin')->nullable();
-            $table->float('total_pagar')->nullable();
+            
+            // Cambiado de float a decimal para máxima precisión en cálculos
+            $table->decimal('total_pagar', 12, 2)->nullable();
             $table->decimal('saldo_pendiente', 12, 2)->default(0);
+            
             $table->string('moneda')->nullable();
             $table->string('lapso_capital')->nullable();
             $table->integer('nro_cuotas')->nullable();
-            $table->integer('tasa')->nullable();
+            
+            // Mantenemos como integer para igualar la tabla plan_pago original
+            $table->integer('tasa')->nullable(); 
+            
             $table->string('tipo_tasa')->default('amortizable');
             $table->integer('estado')->default(1);
             $table->integer('desembolso')->default(1);
             $table->integer('pago_administrativo')->default(1);
-            $table->unsignedBigInteger('id_solicitud')->nullable(); // relación al plan original
+            
+            $table->unsignedBigInteger('id_solicitud')->nullable(); 
             $table->unsignedBigInteger('id_plan_aux')->default(0);
 
-            // Relación con solicitud_respaldo
+            // --- Relación con la tabla solicitud_respaldo ---
             $table->unsignedBigInteger('solicitud_respaldo_id');
             $table->foreign('solicitud_respaldo_id')->references('id')->on('solicitud_respaldo');
 
-            // Información de auditoría
-            $table->unsignedBigInteger('plan_pago_original_id'); // ID del plan_pago original
+            // --- Información de auditoría ---
+            $table->unsignedBigInteger('plan_pago_original_id'); 
             $table->string('accion'); // create, update, delete
+            
             $table->unsignedBigInteger('usuario_accion')->nullable();
             $table->foreign('usuario_accion')->references('id')->on('users')->onDelete('set null');
+            
             $table->ipAddress('ip_address')->nullable();
             $table->text('user_agent')->nullable();
